Rename fetcher to fetch to ensure backward compatibility
Only used fetcher while I was prototyping so I could run old and new at the same time.
