+++ ukat/data/__init__.py
@@ -1,1 +1,1 @@
-from . import fetch, fetcher
+from . import fetch
